requirements.txt
flask
flask_cors
transformers
torch
Pillow
fastapi